Sessions number is an integer.
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@25498 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ protofilter/hier/usr/lib/python2.5/reports/node/untangle_node_protofilter.py
@@ -146,9 +146,9 @@ class ProtocolsHighlight(Highlight):
         one_week = DateFromMx(end_date - mx.DateTime.DateTimeDelta(report_days))
 
         query = """\
-SELECT COALESCE(SUM(new_sessions),0) AS sessions,
-       COALESCE(sum(CASE WHEN NULLIF(pf_protocol,'') IS NULL THEN 0 ELSE 1 END), 0) AS protocols,
-       COALESCE(sum(pf_blocks), 0) AS blocks
+SELECT COALESCE(SUM(new_sessions),0)::int AS sessions,
+       COALESCE(sum(CASE WHEN NULLIF(pf_protocol,'') IS NULL THEN 0 ELSE 1 END), 0)::int AS protocols,
+       COALESCE(sum(pf_blocks), 0)::int AS blocks
 FROM reports.session_totals
 WHERE trunc_time >= %s AND trunc_time < %s"""
